requirements.txt
aiogram>=3.6.0
gspread>=6.0.0
google-auth>=2.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0
asyncpg>=0.29
